requirements.txt
streamlit
langchain_groq
python-dotenv
langchain
PyPDF2